OPHJOD-1437: Optimize test execution
- Consolidate ApplicationTest and ErrorHandlingTest
- Reuse DB container to speed up service tests
- Introduce support for parallel tests.
