+++ src/test/java/fi/okm/jod/yksilo/service/KoulutusServiceTest.java
@@ -32,8 +32,9 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Import;
 import org.springframework.test.context.jdbc.Sql;
+import org.springframework.test.context.jdbc.Sql.ExecutionPhase;
 
-@Sql("/data/osaaminen.sql")
+@Sql(value = "/data/osaaminen.sql", executionPhase = ExecutionPhase.BEFORE_TEST_CLASS)
 @Import({KoulutusService.class, YksilonOsaaminenService.class})
 class KoulutusServiceTest extends AbstractServiceTest {
