+++ src/test/java/fi/okm/jod/yksilo/service/MuuOsaaminenServiceTest.java
@@ -27,8 +27,9 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Import;
 import org.springframework.test.context.jdbc.Sql;
+import org.springframework.test.context.jdbc.Sql.ExecutionPhase;
 
-@Sql("/data/osaaminen.sql")
+@Sql(value = "/data/osaaminen.sql", executionPhase = ExecutionPhase.BEFORE_TEST_CLASS)
 @Import({MuuOsaaminenService.class, YksilonOsaaminenService.class})
 class MuuOsaaminenServiceTest extends AbstractServiceTest {
